requirements.txt
discord.py
pyaudio
requests
openai